create docker compose for database and api
