+++ src/fastapi/README.md
@@ -1,0 +1,3 @@
+# Plant Detector Project
+
+This project uses FastAPI, TensorFlow/Keras, and OAuth2.0 for plant detection.
