--- src/fastapi/app/prediction_management/api_request_log.py
@@ -1,52 +1,0 @@
-from fastapi import APIRouter, Depends, HTTPException, status
-from sqlalchemy.orm import Session
-from typing import List
-
-from app.schemas.api_request_log import APIRequestLogBase, APIRequestLogResponse
-from app.schemas.auth_schema import UserBase
-from app.database.db import get_db
-from app.auth.auth_utils import get_current_admin_user, get_current_user
-
-router = APIRouter()
-
-# Create a new APIRequestLogBase (Typically handled internally)
-@router.post("/api_request_logs/", response_model=APIRequestLogResponse)
-def create_api_request_log(
-     api_request_log: APIRequestLogBase, 
-     db: Session = Depends(get_db),
-     current_user: UserBase = Depends(get_current_user)
-     ):
-     db_api_request_log = APIRequestLogBase(**api_request_log.dict(), user_id=current_user.user_id)
-     db.add(db_api_request_log)
-     db.commit()
-     db.refresh(db_api_request_log)
-     return db_api_request_log
-
-# Get all APIRequestLogs (Admin only)
-@router.get("/api_request_logs/", response_model=List[APIRequestLogResponse])
-def read_api_request_logs(
-     skip: int = 0, limit: int = 100, 
-     db: Session = Depends(get_db),
-     current_user: UserBase = Depends(get_current_admin_user)
-     ):
-     return db.query(APIRequestLogBase).offset(skip).limit(limit).all()
-
-# Get APIRequestLogs for the current UserBase
-@router.get("/api_request_logs/me", response_model=List[APIRequestLogResponse])
-def read_user_api_request_logs(
-     db: Session = Depends(get_db),
-     current_user: UserBase = Depends(get_current_user)
-     ):
-     return db.query(APIRequestLogBase).filter(APIRequestLogBase.user_id == current_user.user_id).all()
-
-# Get a specific APIRequestLogBase by ID (UserBase can access their own, Admin can access all)
-@router.get("/api_request_logs/{request_id}", response_model=APIRequestLogResponse)
-def read_api_request_log(
-     request_id: int, 
-     db: Session = Depends(get_db),
-     current_user: UserBase = Depends(get_current_user)
-     ):
-     db_api_request_log = db.query(APIRequestLogBase).filter(APIRequestLogBase.request_id == request_id).first()
-     if db_api_request_log is None or (db_api_request_log.user_id != current_user.user_id and not current_user.is_admin):
-          raise HTTPException(status_code=404, detail="APIRequestLogBase not found")
-     return db_api_request_log
