+++ src/fastapi/app/utils/model.py
@@ -1,3 +1,3 @@
 import tensorflow as tf
 
-model = tf.keras.models.load_model('../../models/TL_180px_32b_20e_model.keras')
+model = tf.keras.models.load_model('models/TL_180px_32b_20e_model.keras')
